Move data_fit tests into loops
